# contracts/openapi/v1-engine.json
{
  "components": {
    "schemas": {
      "ActiveProviderRequest": {
        "properties": {
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          }
        },
        "required": [
          "provider_id"
        ],
        "title": "ActiveProviderRequest",
        "type": "object"
      },
...
      "ChatMessage": {
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "system",
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
...
      "CompleteRequest": {
        "properties": {
          "connection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          }
        },
        "required": [
          "messages"
        ],
        "title": "CompleteRequest",
        "type": "object"
      },
...
      "CredentialsRequest": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "default_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Model"
          }
        },
        "required": [
          "api_key"
        ],
        "title": "CredentialsRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterpretRequest": {
        "properties": {
          "provider_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "roster": {
            "items": {
              "$ref": "#/components/schemas/RosterEntry"
            },
            "title": "Roster",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "InterpretRequest",
        "type": "object"
      },
...
      "OllamaPresetRequest": {
        "properties": {
          "model": {
            "default": "qwen3.5:4b",
            "title": "Model",
            "type": "string"
          }
        },
        "title": "OllamaPresetRequest",
        "type": "object"
      },
...
      "RosterEntry": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "default": "person",
            "enum": [
              "person",
              "agent",
              "other"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "RosterEntry",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local Homun 2 engine — domain F2 over SQLite + HTTP.",
    "title": "Homun Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models/chat": {
      "post": {
        "description": "Alias of /complete — ModelPort chat entrypoint.",
        "operationId": "chat_complete_v1_models_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Chat Complete V1 Models Chat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Complete",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/chat/stream": {
      "post": {
        "description": "SSE token stream for ModelPort chat (live chunks then result).",
        "operationId": "chat_stream_v1_models_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat Stream",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/complete": {
      "post": {
        "operationId": "complete_v1_models_complete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Complete V1 Models Complete Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete",
        "tags": [
          "models"
        ]
      }
    },
...
    "/v1/models/interpret": {
      "post": {
        "operationId": "interpret_message_v1_models_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Interpret Message V1 Models Interpret Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Interpret Message",
        "tags": [
          "models"
        ]
      }
    },
...
    "/v1/models/providers/active": {
      "post": {
        "operationId": "set_active_provider_v1_models_providers_active_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActiveProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Active Provider V1 Models Providers Active Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Active Provider",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/providers/openai_compatible/credentials": {
      "delete": {
        "operationId": "clear_openai_credentials_v1_models_providers_openai_compatible_credentials_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clear Openai Credentials V1 Models Providers Openai Compatible Credentials Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Clear Openai Credentials",
        "tags": [
          "models"
        ]
      },
      "post": {
        "operationId": "set_openai_credentials_v1_models_providers_openai_compatible_credentials_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Openai Credentials V1 Models Providers Openai Compatible Credentials Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Openai Credentials",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/providers/openai_compatible/ollama_preset": {
      "post": {
        "operationId": "apply_ollama_preset_v1_models_providers_openai_compatible_ollama_preset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OllamaPresetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Apply Ollama Preset V1 Models Providers Openai Compatible Ollama Preset Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Ollama Preset",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/providers/{provider_id}/verify": {
      "post": {
        "operationId": "verify_provider_v1_models_providers__provider_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verify Provider V1 Models Providers  Provider Id  Verify Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Provider",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/usage": {
      "get": {
        "operationId": "list_usage_v1_models_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Usage V1 Models Usage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Usage",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/models/usage-attempts": {
      "get": {
        "description": "F3.5: interpret/model attempts with command context; tokens never invented as zero.",
        "operationId": "list_usage_attempts_v1_models_usage_attempts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Usage Attempts V1 Models Usage Attempts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Usage Attempts",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}